+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # koda-times change Logs
 
 
-# 1.0.1
+# 1.1.0
 
 *   Add arithmatic operation for Period, Instace, LocalDateTime, LocalDate, LocalTime
 *   Remove DurationBuilder class
+* Upgrade kotlin library to 1.1 
 
 # 1.0.0